pyproject.toml
[project]
name = "mlops"
version = "0.1.0"
description = ""
authors = [
    {name = "Max32x",email = "maxime-32@live.fr"}
]
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "flask (>=3.1.0,<4.0.0)",
    "scikit-learn (>=1.6.1,<2.0.0)",
    "pandas (>=2.2.3,<3.0.0)",
    "pandas-gbq (>=0.26.1,<0.27.0)",
    "google-cloud-storage (>=2.19.0,<3.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
